remove useless folder
