+++ tests/e2e/protectData.test.ts
@@ -2,9 +2,9 @@ import fsPromises from 'fs/promises';
 import path from 'path';
 import { beforeEach, describe, expect, it } from '@jest/globals';
 import { HDNodeWallet, Wallet } from 'ethers';
-import { IExecDataProtector, getWeb3Provider } from '../../../src/index.js';
-import { ValidationError, WorkflowError } from '../../../src/utils/errors.js';
-import { MAX_EXPECTED_BLOCKTIME } from '../../test-utils.js';
+import { IExecDataProtector, getWeb3Provider } from '../../src/index.js';
+import { ValidationError, WorkflowError } from '../../src/utils/errors.js';
+import { MAX_EXPECTED_BLOCKTIME } from '../test-utils.js';
 
 describe('dataProtector.protectData()', () => {
   let dataProtector: IExecDataProtector;
